Fix embedding has_many with ordered associations
Previously this would break, because the ORDER BY did not match the
generated GROUP BY clause. This change moves the ORDER BY clause into
the array_agg aggregation keeping the correct order of ids.

There doesn't seem to be a nice way to include an ORDER BY into a named
function node, so a hack is used by first building a select node,
stripping the SELECT keyword from the sql and then passing it as a literal.

Two cases are handled:

* `order` in `default_scope` of associated model
* `order` in custom `has_many` association scope