Don't ignore arel bind values in filter query
This fixes missing bind values when conditions are on the arel select
manager, for example join conditions. This is only relevant for AR 4.2+,
because older versions don't use prepared statements for these queries.

There's also a small fix for AR 4.0 to avoid ambigous column names,
because in this version `select(:id)` does not add the table name.