Add support for single record serialization
This will handle serialization of single records as used for example in
show actions using the postres array serializer. To be able to do this,
the record is wrapped in a relation using its primary key for lookup.

It's also possible to use this feature with a relation by passing the
`root: 'singular_name'` and `single_record: true` options to render or
the serializer, but you have to ensure that the relation only returns a
single record, for example by using `.limit(1)`.

The single record serialization format does use a singular root key name
for the main resource and does not wrap the object in an array:

Collection: `{"notes":[{"id":1,"tag_ids":[1]}],"tags":[...]}`
Single Record: `{"note":{"id":1,"tag_ids":[1]},"tags":[...]}`

This adds a patch to ActiveModel::Serializer.build_json which adds the
required wrapping in a relation.

Note that this method will load and instantiate the record and it's
serializer once and then reload it using the primary key in a relation.
To avoid this you can skip the find call in the controller and replace
it with a relation and pass `:root` and `:single_record` along with the
singular relation in the call to `render`.